[feat/#187] 내부 컴포넌트 파일 분리
